All enums migrated to enum class. Added type cast in all places where used. Added projection reader. In reader form added preview of file if can be previewed.

+++ gui/SaveShortPathForm.cpp
@@ -163,10 +163,10 @@ void SaveShortPathForm::saveShortPath()
 
     QByteArray fileNameLatin = fileName.toLatin1();
 
-    unsigned options = WriterBase::Option::NONE;
+    unsigned options = (unsigned) WriterBase::Option::NONE;
     if (ui->printIndentsCheck->isChecked())
     {
-        options = options | WriterBase::Option::PRINT_INDENTS;
+        options = options | (unsigned) WriterBase::Option::PRINT_INDENTS;
     }
     NodeIdDeque nodes;
     nodes.push_back(nodeId);
